requirements.txt
apache-airflow-providers-amazon>=7.4.0 